requirements.txt
asgiref==3.8.1
bcrypt==4.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
dj-database-url==2.3.0
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
pillow==10.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2025.1
redis==5.0.3
sqlparse==0.5.3
typing_extensions==4.12.2
